+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `TryFirst` extension for `IEnumerable`
 - `TryLast` extension for `IEnumerable`
 - `TryElementAt` extension for `IEnumerable`
+- `Collect` extension for `IEnumerable`
 ## 2.0.1
 ### Fixed
 - `FromNullable` method did not work with some types, like `string` but now it  does